Create adaptable base class
Allows reuse of adapter logic
